Initial commit to reset master branch
If you need any older history, see the dev branch version. I screwed up early on by putting all my work into master branch which exposed all the WIP (work in progress) code to anyone checking it out. Now the working branch is dev, and master is "stable" for release, and we will be using a more traditional gitflow to manage the project with features branches off dev, and performing regular merges. As such, this master branch was re-created as an orphan with history being reset to here.