Updates PHPGoodies::resourceInfo method to accept directory-loading
Directory loading allows us to specify a class by the directory name that it lives in such that the class code lives in implementation.php within that directory. This allows us to place everything needed for that class into the directory (test, example, etc) so that all the bits for that Goody are kept together. In the next pass we will enhance the contents of such directories so that a given Goody can record dependencies on other Goodies which will allow us to build a dependency tree and ultimately allow a given application to pull in only those Goodies that it needs along with all the dependencies of those Goodies automatically (yay for sub-repository dependency management!)