Adds seed class for Lib.File.Pdf goody
Needs implementation to be filled in. Maybe we should create an interface for our various Lib.File classes that all have parse/create/update, etc. public interface methods (in addition to their file type specific operations needed to work with them) to standardize the way in which we generally work with structured files...?)