Adds the class shell for PHPDeps Utility goodie
We want to examine each PHP file provided, change to its directory, and follow relative paths as well as partial paths from the script execution directory as well as each file's directory to see whatever other PHP files we can find... and recurse on each of them. This will require looking at requires and includes and doing some file finds on the filesystem - it might be easiest if we build a filesystem tree in memory in advance and then scan that from in-memory instead of repeatedly doing finds on the filesystem.