Adds a Document class for JSON:API response documents
This is a structured class that serves as the foundation layer for all JSON_API responses. It's a little manual as it is currently designed simply to use an OOP approach to building a response data object, but we can see about making convenience improvements once we get the base working and tested.